requirements.txt
streamlit>=1.28.0
pydub>=0.25.1
PyPDF2>=3.0.1
python-docx>=0.8.11
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
pyttsx3>=2.90
gTTS>=2.3.0
pygame>=2.5.0